Add one-file-script generator
Copied from the [todo.txt-mit project][1], the `generate_script`-script
concatenates all the source files into a single executable (which still
needs to have the executable flag set afterward). This works as an
alternative way of providing an executable for todo.sh to run (instead
of symlinking).

Adjust the README to reflect this addition.

  [1]: https://github.com/linduxed/todo.txt-mit