Add padding to run commands through Executable
Since todo.sh runs add-ons in such a way that the arguments to the
add-ons can be found in ARGV[1] and later, we need to pad the commands
that are run with an arbitrary word.

This does not apply to one case (usage message from `todo.sh help`), but
we can disregard this for now.