Make Executable helper accept test fixtures
To allow for testing the add-on with custom TODO_FILE instances, while
not mocking anything internally, we need to pass a location for an
alternate TODO_FILE through ENV variables.

Extend the Executable helper and add another helper method
(`with_todo_file`) to allow for creating a temporary TODO_FILE and
injecting the path to it.

As a minor note: the invocation of `gsub!` is there to allow for writing
heredocs describing TODOs that are indented, instead of having to
deindent them to the start of the line.