Add failing acceptance test for linking card
The acceptance test added in this patch checks that three things happen
after invoking "todo.sh trello link 123 https://url.to/some/card":

1. The add-on output includes a "Trello tag".
2. The URL has been added to the database, under above mentioned tag.
3. The task indicated by the number has been modified in the TODO_FILE,
   where the tag and a short indicator prefix has been added.

To allow a database fixture to be provided to the add-on, extend the
`with_todo_file` helper and rename it to `with_todo_file_and_database`.

As a minor note: in `with_todo_file_and_database` the passed in `todos`
have their leading whitespace removed with `gsub!`, to allow for writing
`todos` in an indented fashion. This is not necessary for the
`database_entries`, since `YAML.safe_load` seems to truncate leading
indentation up to a global minimum (i.e. if all lines have at least
2 spaces of indentation, then 2 spaces are removed from each line).