Add first version of `link` functionality
The code in this patch is primarily intended to make the red acceptance
test pass. The code will probably need refactoring, error handling and
unit tests.

The general idea of `link` is that a record is added to the database,
and a chosen TODO is updated to contain a tag.

The database is a YAML file, where the top-level keys are the tags that
link the Trello-related data in the database with local TODOs in the
TODO_FILE.

The tags are strings consisting of the capital letters A-Z, with the
"first" one being "A". New tags are obtained by applying String#succ!
(starting with "A") repeatedly until an available tag is found.

The TodoFileMutator#add_tag_to_todo method added in this patch adds the
tag by simply appending it to the end of the chosen TODO. I think it may
look better if trailing contexts ("@foo") and projects ("+bar") were
placed after the tag, but this would result in more code and would be
more error prone: I can't assume that all TODOs have their contexts and
projects at the end.