Refactor Database and create Database::Record
The changes in this patch were prompted by the fact that CLI shouldn't
need to know about the internals database records to be able to obtain
their tags. Add a Database::Record class which provides readers for the
various attributes that records can have.

Additionally, refactor the internals of the Database class to operate on
lists of Database::Record.