Add dedicated class for TODO_FILE interactions
Currently the logic for finding a TODO exists and strewn out code in
a branch in the CLI#run method. Additionally, mutating the TODO_FILE has
its own class (`TodoFileMutator`) when actions not involving mutation
are also done (just finding and reading a TODO).

Place both of these behaviours in one class.

Additionally, have the string output from the case branch in CLI#run
read the modified TODO instead of outputting a composed string that
might not correspond to the state of the TODO after applying
`TodoFile.add_tag_to_todo`.