Ensure that Trello card database exists
Since the card database is integral to the correct working of the
add-on, I think it makes sense to always create it if it's missing.

The alternative would be to require the user create the database with
`touch` prior to using the add-on, which seems like an unnecessary extra
burden.