Refactor TodoFile#add_tag_to_todo and #find_todo
Extract the notion of reading a specific TODO into a private method,
since this is shared between TodoFile#add_tag_to_todo and
TodoFile#find_todo.

Additionally, extract a private method for writing to the TODO_FILE, for
clarity's sake.

Finally, it should be noted that this patch increases the amount of
times the entire file is read with File.readlines. While accessing the
disk unnecessarily bothers me somewhat, in this case the alternative
would be to make the code more cluttered with file handling. For now,
I'll run with the less performant option; I'll likely change this later.